SSVC commit to Git for Github. Actual history in .simpsvc dir
SSVC is a simple source version control logic, which allows one
to track changes in a structured manner.

As part of keeping a copy of the same in github. I have created
a dummy git repo for the same.

However actual history of development of this version control
system can be looked at by using hkvc-simple-svc.py on itself.
The history as well as the repository is in .simpsvc directory.

ssvc was developed and tracked using itself, from a relatively
early stage.

This was done to work with some projects I maintained in a sd card,
while using a chromebook. The plan 9 filesystem exposed by chrome os
was not compatible with .git, so I created this just for some fun,
as well as to track those projects.

NOTE: A post facto git commit of things done previously